상품상세 리뷰보기, 리뷰작성, 리뷰삭제 기능 완료, 마켓글삭제기능 확인취소로 변경완료

+++ src/main/webapp/WEB-INF/client/market/market-read.jsp
@@ -103,8 +103,12 @@
 <jsp:include page="../include/footer.jsp" />
 <script>
 function deleteMarket() {
-	alert("정말로 삭제 하시겠습니까? ");
-	document.location.href="market-delete.do?marketTbNo=" + ${market.marketTbNo};
+	var result = confirm("정말로 삭제 하시겠습니까? ");
+	if(result){
+		document.location.href="market-delete.do?marketTbNo=" + ${market.marketTbNo};		
+	} else {
+		return false;
+	}
 }
 
 //start jqury
